Update type extractors
We add a RAI TypeExtractor to handle some cases, however it may not be
as helpful as one might think. The TypeService reifies the type of an
object by first obtaining its class and then uses the TypeExtractors to
attempt to reify the type variables. Thus the RAI TypeExtractor can be
used to identify type variables in subclasses of the interface (i.e.
    ArrayImg, CellImg, etc).

We should ask, however, how many subclasses of RAI actually only have
the one type variable? Certainly the View types do (i.e. ShuffledView),
    and ListImg does as well. But many others (especially the ones used
	often, such as ArrayImg) have more type variables. The RAI
    TypeExtractor will not be able to reify those other type variables,
    which can be problematic in matching. Thus we cannot get rid of
    TypeExtractors that are specific to types like ArrayImg, and we will
    probably have to add other TypeExtractors in the future as the need
    arises.

We also make a couple of changes to the scijava-ops TypeExtractors, most
notably deleting the GenericTypedTypeExtractor. We delete it because
there is a check in TypeSerivce.reify() that uses instanceof() to
determine whether the Object being reified is a GenericTyped and
handling it from there. This makes the GenericTypedTypeExtractor
unreachable.