Union org.scijava.plugin.Plugin files for tests
For the longest time we could not figure out why the Ops in the main
package were not being called in the tests of scijava-ops. I first
thought that it was a discovery issue, however that quickly turned out
not to be the issue; I found this out by adding a second
module-info.java to the test package, and while this broke Eclipse it
showed (in the org.scijava.plugin.Plugin file) that both packages WERE
being seen during test compilation.

After a TON of further digging, I finally found PatchedModuleReader.findResources(). It turns out that,
given a name, we are only able to return a single URL ending in that
name PER MODULE. But, because during test compilation the test classes
are patched into the MAIN classes, we needed to see BOTH of the
org.scijava.plugin.Plugin files in order to use both main and test
package Ops. Furthermore, PatchedModuleReader.findResources() always
looks for the resource in its patches before considering its own
resources, we know that we will always find the resource in the patch
(i.e. the org.scijava.plugin.Plugin in target/test-classes).

With this in mind, we decided that we could either rename one of the two
org.scijava.plugin.Plugin files or we could paste the main package data
into the test package data. We ended up choosing the latter, since it
was easier to do and only required a script to be run. It is a bit
disappointing that we have data in the test package that refers to the
main package, but maybe this is okay since nobody will ever want the
test package without the main package as well.