Migrate most tests to use JUnit 5
JUnit 4 was nice and simple, however we noticed that the complexity of
our tests in ImageJ Ops (with multiple hierarchies of Abstract test
classes) caused Eclipse troubles when it came to initializing our setup
methods. This would cause errors in Eclipse (and in Eclipse only).

We hoped that an update to JUnit 5 would solve this (and helpfully be
more friendly to a modular project). This commit updates all of our
tests in imagej-ops2 to use JUnit 5. It also removes the dependency on
JUnit4.

One wrinkle: Maven is unable to build Ops2 without the presence of at
least one class in the net.imagej.ops2 package. It would be great if we
did not need a Dummy class; see next commit for a better placeholder.