OpMonitorTest: fix asserts
There are really two fixes going on here:

1. Use JUnit Assert methods instead of Java's assert
2. Assert that the current progress of the OpMonitor is 1 in
testProgress(), NOT the progress variable. It is possible that the
progress variable was not updated before the while loop ended (this is
actually pretty likely, and we were probably passing due to rounding).
Thus the progress variable could be off (and Travis CI found that it
was). By checking monitor.getProgress() instead, we can obtain the
actual value of the progress after the Future reports as done.