Catch errors thrown by getExactSuperType
This usually implies some funny business going on (which we do not want
to support, so we have checkGenericAssignability return false upon
catching an error).

The example driving this change can be realized by the following
structure.

public <I extends Number> void Foo () {
  class Bar implements Function<I, Double> { // apply method here}
  Nil<Function<Double, Double>> barFunc = new Nil<>(){};
  MatchingUtils.checkGenericAssignability(Foo.class, barFunc.getType(),
      false);
}

This will cause GenTyRef to throw an AssertionError because Bar does not
declare the TypeVariable it uses, and this is the error we are hoping to
catch.

We decide to return false in these situations because we DO NOT KNOW if
we can say

Function<Double, Double> barObj = new Bar();

And, if we were to try within the method, we would get a compile error.
Since we cannot do it here (and in general it depends on the context in
which you try to create that Bar object), we return false for
consistency.