inferTypeVariables: clean up errors
It is ideal to explain why the method fails when it does, in fact, fail.
This change gives more clarity as to why we are failing, since there are
two ways we can fail here:

1. TypeVarAssigns is null: Why are you calling this method if you don't
have a Map in which you are storing the mappings??

2. types and inferFroms have a mismatch in length. In this case it is
unclear which type in types maps to which type in inferFroms.