inferTypeVariables: add typeVarAssigns disclaimer
The user should needs to understand that the mappings are probably going
to be incorrect when an error is thrown. Often an error is thrown
when the caller wants to map a TypeVariable to two different types. In
this scenario, the TypeVariable will be mapped to only one of those two
types in the map. But this is obviously incorrect, so the user should be
aware of that when an error is thrown.