TypesTest: correct isAssignableT test
We know that the following is compile-error free:

public void barFunc() {
    List<Number> list = new ArrayList<>();
    list.add(1.);
    list.add(1l);
    fooFunc(list);
}

public <T extends Number> void fooFunc(List<T> list){
}

This would suggest that we can implicitly cast a List<Number> to a
List<T>, and that by extension isAssignable(listNumber, listT) should
return true.

A similar argument follows for the next two calls to isAssignable (but
not for the last call in the block since the analogous code for that
call results in a compiler error).