TypeInferenceException: should be unchecked
This provides two benefits:

1) This prevents the more abstract, high-level methods within
MatchingUtils from knowing too much about the implementation in the
low-level methods (by declaring TypeInferenceExceptions all over the
place). This also makes the code cleaner :)

2) This prevents the leakage of TypeInferenceException into other
classes (specifically Types). It would be nice if, for example,
TypeMapping could throw TypeInferenceExceptions, however this requires
TypeVarAssigns to throw them as well (which is a problem for Types).
This is less of a concern, however still a concern.