Improve retyping for edge case scenarios
In rare edge cases, functional interfaces don't declare their abstract
method. This might happen for a convenience interface like:

interface AllSameFunction<T> extends BiFunction<T, T, T> {}

such an interface does not have a declared abstract method, so we have
to find it via its superinterface BiFunction. The method we get out of
BiFunction, however, will not know of the mapping of the Ts in
AllSameFunction to the T, U, R of BiFunction; this is a bug in the
method.

This commit adds logic that determines the mappings of the type
variables of interfaces like AllSameFunction to the type variables of
BiFunction, so that Ops implementing AllSameFunction can be retyped.