Improve determination of suitable Op wrapper
When Op wrappers exist for both a class and one of its superclasses, Op
wrapping would fail as multiple wrappers would be suitable. The Op
wrapping logic was not smart enough to filter these matches in any way.

This commit adds a method to filter superclasses out of the list of
suitable wrappers. Since these wrappers also preserve the exact
supertype of the Op w.r.t. the wrapper class, it is desirable to have
subclasses declare their own wrappers, and this fix prevents multiple
suitable wrappers from preventing a wrapper match (so long as all
classes are related).

This commit also fixes a bug in the error message when multiple wrappers
would match. The error was designed to print the classes of each wrapper
that matched, but would instead vomit out the classes of all known
wrappers. This bug has also been fixed.