Stream methods resembling the functional method
We want to leave open the possibility for @Optional annotations to be
declared on the interface as well as on the Op implementation. This
means we have to check multiple Methods for the Optional annotation.
This could be a source of bugs, but for now we check any method with the
same name as the functional method, having the same number of
parameters.