Consolidate create packages
The current situation is untenable since Creators depends on the
net.imagej.ops2.create.kernel and net.imagej.ops2.create.img packages.
Ideally, we'd split out the image- and kernel create Ops into their own
packages like we had, but the kernel and image creator Op fields rely on
the (non-static) Op fields in the main package. Therefore if we moved
these Ops to the subpackages, we would have to duplicate code. To
prevent this duplication, we shall just consolidate to one package.