Prioritize a more powerful version of setHints
Previously we prioritized a setHints taking only the whole Hint. Now we
prioritize a Hint composed of the String hintType and the Object value.
We also add a getDelimiter method to provide an
implementation-independent way to get the delimiter. In this way we no
longer tie ourselves to a single delimiter.