Remove Ops connotation from Scijava Discovery
It is unfortunate to make this change on this branch, but after writing
SciJava Ops ServiceLoader, it became clear that SciJava Ops Discovery
has nothing to do with Ops. Thus it becomes SciJava Discovery (without
the "Ops). SciJava Ops ServiceLoader, however, needs the Ops name
because of its module-info.java