Create a proper "filter.fft" function op
Running `out = ops.op("filter.fft").arity1().input(img).apply()` returns
a blank image. This is because the wrong image dimensions are being used
for the FFT result. This is an example of when adaptation fails. SciJava
Ops adapts the FFT Op into a function and assumes that the output image
dimensions are the same as the input. Generally this is a safe
assumption, but this is an example where the output image dimensions
are in fact different. This commit adds a new Op that depends on
`FFTMethodsOpF` and returns the FFT result as a ComplexFloatType. This
is consistent with the old imagej-ops usage.